@@ -19,7 +19,7 @@ void main() {
     expect(
       find.byWidgetPredicate(
         (Widget widget) => widget is Text &&
-                           widget.data.startsWith('Running on:'),
+                           (widget.data as String).startsWith('Running on:'),
       ),
       findsOneWidget,
     );
