@@ -33,7 +33,7 @@ class PdftronFlutter {
   }
 
   static Future<void> openDocument(String document,
-      {String password, Config config}) {
+      {String? password, Config? config}) {
     return _channel.invokeMethod(Functions.openDocument, <String, dynamic>{
       Parameters.document: document,
       Parameters.password: password,
@@ -46,7 +46,7 @@ class PdftronFlutter {
         Functions.importAnnotations, <String, dynamic>{Parameters.xfdf: xfdf});
   }
 
-  static Future<String> exportAnnotations(List<Annot> annotationList) async {
+  static Future<String?> exportAnnotations(List<Annot>? annotationList) async {
     if (annotationList == null) {
       return _channel.invokeMethod(Functions.exportAnnotations);
     } else {
@@ -99,19 +99,19 @@ class PdftronFlutter {
         <String, dynamic>{Parameters.bookmarkJson: bookmarkJson});
   }
 
-  static Future<String> saveDocument() {
+  static Future<String?> saveDocument() {
     return _channel.invokeMethod(Functions.saveDocument);
   }
 
-  static Future<bool> commitTool() {
+  static Future<bool?> commitTool() {
     return _channel.invokeMethod(Functions.commitTool);
   }
 
-  static Future<int> getPageCount() {
+  static Future<int?> getPageCount() {
     return _channel.invokeMethod(Functions.getPageCount);
   }
 
-  static Future<bool> handleBackButton() {
+  static Future<bool?> handleBackButton() {
     return _channel.invokeMethod(Functions.handleBackButton);
   }
 
@@ -127,12 +127,12 @@ class PdftronFlutter {
     return pageRotation;
   }
 
-  static Future<bool> setCurrentPage(int pageNumber) {
+  static Future<bool?> setCurrentPage(int pageNumber) {
     return _channel.invokeMethod(Functions.setCurrentPage,
         <String, dynamic>{Parameters.pageNumber: pageNumber});
   }
 
-  static Future<String> getDocumentPath() {
+  static Future<String?> getDocumentPath() {
     return _channel.invokeMethod(Functions.getDocumentPath);
   }
 
